requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
blis==0.2.4
certifi==2019.3.9
chardet==3.0.4
ConfigArgParse==0.14.0
cycler==0.10.0
cymem==2.0.2
fasttext==0.8.22
future==0.17.1
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
idna==2.8
jsonschema==3.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
mock==3.0.5
murmurhash==1.0.2
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
plac==0.9.6
preshed==2.0.1
protobuf==3.8.0
pybind11==2.2.4
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
spacy==2.1.4
srsly==0.0.5
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-hub==0.4.0
termcolor==1.1.0
thinc==7.0.4
torch==1.1.0
torchtext==0.4.0
torchvision==0.3.0
tqdm==4.30.0
urllib3==1.25.3
wasabi==0.2.2
Werkzeug==0.15.4
